PM inbox

+++ starter/src/main/java/com/github/vimboard/mapper/NoticeboardMapper.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Mapper
 public interface NoticeboardMapper {
 
-    List<Noticeboard> preview(long limit);
+    List<Noticeboard> preview(int limit);
 }
